# perf: warm expensive init code at snapshot time