WIP -- 10% startup time improvement:

```
14:04 $ hyperfine -S none --warmup 10 '/tmp/deno run /tmp/empty.js' 'target/release/deno run /tmp/empty.js'
Benchmark 1: /tmp/deno run /tmp/empty.js
  Time (mean ± σ):      17.6 ms ±   5.3 ms    [User: 11.3 ms, System: 4.2 ms]
  Range (min … max):    14.9 ms …  84.3 ms    175 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 2: target/release/deno run /tmp/empty.js
  Time (mean ± σ):      15.9 ms ±   0.8 ms    [User: 10.5 ms, System: 3.9 ms]
  Range (min … max):    14.3 ms …  18.4 ms    183 runs
 
Summary
  'target/release/deno run /tmp/empty.js' ran
    1.10 ± 0.34 times faster than '/tmp/deno run /tmp/empty.js'
✔ ~/Documents/github/deno/deno [main ↓·2|✚ 4…4⚑ 23] 
```