@mmastrac 2024-03-16
@bartlomieju Confirmed that `__runtime_js_sources` continues to work with this patch.